+++ paleoDiv/R/paleoDiv.R
@@ -195,7 +195,7 @@ return(y_)
 #'Generate a violin plot
 #'
 #' @param x Variable for which to plot violin.
-#' @param pos Position at which to place violin in the axis perpendicular to x
+#' @param pos Position at which to place violin in the axis perpendicular to x. Defaults to 0
 #' @param x2 Optional variable to use instead of x as input variable for the violin plot. If x2 is set, the function (default: density()) used to calculate the plotting statistic is run on x2 instead of x, but the results are plotted at the corresponding x values.
 #' @param stat The plotting statistic. Details to the density() function, as in a standard violin plot, but can be overridden with another function that can take x or x2 as its first argument. Stat can also be a numeric vector of the same length as x, in which case the values in this vectors are used instead of the function output and plotted against x as an independent variable.
 #' @param dscale The scale to apply to the values for density (or another plotting statistic). Defaults to 1, but adjustment may be needed depending on the scale of the plot the violin is to be added to.
@@ -221,7 +221,7 @@ return(y_)
 #' viol(c(1:10), width=9, stat=rmean, pos=0, add=FALSE)
 #' viol(c(1:10), stat=c(11:20), pos=0, add=FALSE)
 
-viol<-function(x, pos, x2=NULL, stat=density, dscale=1, cutoff=range(x), horiz=TRUE, add=TRUE,lim=cutoff,xlab="",ylab="", fill="grey", col="black", lwd=1, lty=1,...){
+viol<-function(x, pos=0, x2=NULL, stat=density, dscale=1, cutoff=range(x), horiz=TRUE, add=TRUE,lim=cutoff,xlab="",ylab="", fill="grey", col="black", lwd=1, lty=1,...){
 #sort ascending
 if(!is.null(x2) & is.numeric(x2) & length(x2)==length(x)){
 x2[order(x)]->x2
@@ -746,8 +746,8 @@ abdistr_<-function(x, table=NULL, ab.val=table$abund_value, smooth=0, max=table$
     which(as.numeric(max)>=x)->b
     intersect(a,b)->id
 
-    #sum(ab.val[id], na.rm=TRUE)->abundance
-    length(id)->abundance
+    sum(ab.val[id], na.rm=TRUE)->abundance
+    #length(id)->abundance
     return(abundance)
     }
 
@@ -1080,6 +1080,7 @@ return(sptab_)
 #' @param occ Either a list()-object containing taxon-range tables for plotting diversity, or a matrix() or data.frame()-object that contains numerical plotting statistics. If the latter is provided, the default use of divdistr_() is overridden and the function will look for a column named "x" and columns matching the phylogeny tip.labels to plot the spindles.
 #' @param stat Plotting statistic to be passed on to viol(). Defaults to use divdistr_().
 #' @param prefix Prefix for taxon-range tables in occ. Defaults to "sptab_"
+#' @param pos Position at which to draw spindles. If NULL (default), then spindles are drawn at c(1:n) where n is the number of taxa in phylo0.
 #' @param ages Optional matrix with lower and upper age limits for each spindle, formatted like the output of tree.ages() (most commonly the same calibration matrix used to time-calibrate the tree)
 #' @param xlimits Limits for plotting the on the x axis.
 #' @param res Temporal resolution of diversity estimation (if occ is a matrix or data.frame containing plotting statistics, this is ignored)
@@ -1095,6 +1096,7 @@ return(sptab_)
 #' @param labels Logical indicating whether to plot tip labels of phylogeny (defaults to TRUE)
 #' @param txt.y y axis alignment of tip labels
 #' @param txt.x x coordinates for plotting tip labels. Can be a single value applicable to all labels, or a vector of the same length as phylo0$tip.label
+#' @param adj.x Numeric value giving alignment on x axis, defaults to 0 (left-aligned) but can also be 0.5 (centered) or 1 (right-aligned).
 #' @param add Logical indicating whether to add to an existing plot, in which case only the spindles are plotted on top of an existing phylogeny, or not, in which case the phylogeny is plotted along with the spindles.
 #' @param tbmar Top and bottom margin around the plot. Numeric of either length 1 or 2
 #' @param smooth Smoothing parameter to be passed on to divdistr_()
@@ -1112,7 +1114,7 @@ return(sptab_)
 #' phylo.spindles(tree_archosauria,occ=archosauria,dscale=0.005,ages=ages_archosauria,txt.x=66)
 #' phylo.spindles(tree_archosauria,occ=diversity_table,dscale=0.005,ages=ages_archosauria,txt.x=66)
 
-phylo.spindles<-function(phylo0, occ, stat=divdistr_, prefix="sptab_", ages=NULL, xlimits=NULL, res=1, weights=1, dscale=0.002, col=add.alpha("black"), fill=col,lwd=1, lty=1, cex.txt=1,col.txt=add.alpha(col,1), axis=TRUE, labels=TRUE, txt.y=0.5,txt.x=NULL, add=FALSE,tbmar=0.2,smooth=0){
+phylo.spindles<-function(phylo0, occ, stat=divdistr_, prefix="sptab_", pos=NULL,ages=NULL, xlimits=NULL, res=1, weights=1, dscale=0.002, col=add.alpha("black"), fill=col,lwd=1, lty=1, cex.txt=1,col.txt=add.alpha(col,1), axis=TRUE, labels=TRUE, txt.y=0.5,txt.x=NULL,adj.x=0, add=FALSE,tbmar=0.2,smooth=0){
 
 if(length(tbmar)==1){tbmar<-rep(tbmar,2)}#if only one value is given for tbmar, duplicate it. Otherwise, first value is bottom, second top
 if(inherits(phylo0,"phylo")){#setting for phylogenetic tree
@@ -1121,7 +1123,8 @@ txt.x<-phylo0$root.time-txt.x
 if(is.null(xlimits)){
 xlimits<-c(round(phylo0$root.time)-1,0)}
 
-}else if(is.character(phylo0)){taxsel<-phylo0#settings for taxonomic list
+}else if(is.character(phylo0)){
+taxsel<-phylo0#settings for taxonomic list
 if(!is.null(ages) & is.null(xlimits)){
 xlimits<-rev(range(ages))}
 
@@ -1129,6 +1132,17 @@ xlimits<-rev(range(ages))}
 
 if(is.null(txt.x)){txt.x<-mean(xlimits)}
 
+#set y positions at which to plot, if unspecified
+if(is.null(pos)){
+pos<-c(1:length(taxsel))
+}
+if(length(pos)<length(taxsel)){
+pos<-c(1:length(taxsel))
+}
+#set y axis adjustment, if unspecified
+if(length(adj.x)<length(taxsel)){
+adj.x<-rep(adj.x,length(taxsel))[1:length(taxsel)]
+}
 
 
 ##generate base plot
@@ -1138,7 +1152,7 @@ if(inherits(phylo0,"phylo")){
 ape::plot.phylo(phylo0,x.lim=-1*(xlimits-phylo0$root.time),align.tip.label=2, label.offset=50,show.tip.label=FALSE, y.lim=c(1-tbmar[1],length(phylo0$tip.label)+tbmar[2]))->plot1
 
 }else{
-plot(NULL, xlim=xlimits,ylim=c(1-tbmar[1], length(taxsel)+tbmar[2]), xlab="",ylab="",axes=FALSE)
+plot(NULL, xlim=xlimits,ylim=c(min(pos)-tbmar[1], max(pos)+tbmar[2]), xlab="",ylab="",axes=FALSE)
 plot1<-NULL
 }
 }else{#if add==FALSE
@@ -1174,14 +1188,12 @@ if(!is.null(ages)){#if age data is provided
         if(nrow(ages)==length(taxsel)){#contingency if rownames cannot be matched, but row numbers can:
         cutoff<-as.numeric(ages[i,])
         }else{
-        cutoff<-range(eval(parse(text=paste0("occ$",prefix,taxsel[i]))), phylo0[,2:3])}
+        cutoff<-range(eval(parse(text=paste0("occ$",prefix,taxsel[i])))[,2:3])}
         }
-}else{#use range of data if no validly formatted ages are provided
-cutoff<-range(eval(parse(text=paste0("occ$",prefix,taxsel[i]))), phylo0[,c("max","min")])###XXX
 }
 
-if(inherits(phylo0,"phylo")){phylo0$root.time-cutoff->cutoff}
-#end spindle limits
+if(inherits(phylo0,"phylo") & exists("cutoff")){phylo0$root.time-cutoff->cutoff}#convert cutoffs to phylo plotting space, if needed
+#end setting of spindle limits
 
 
 ##vary colors, if desired
@@ -1206,7 +1218,7 @@ if(length(w)!=length(seq(min(plot1$x.lim),max(plot1$x.lim),abs(res)))){stop("wei
 
 #prepare data for plotting
 
-if(!is.matrix(occ) & !is.data.frame(occ)){
+if(!is.matrix(occ) & !is.data.frame(occ)){## if list of multiple dataframes is given, convert data.frame for use with divdistr_() or abdistr_()
 
 itable<-occ[[paste0(prefix,taxsel[i])]]
 
@@ -1226,15 +1238,18 @@ plotx<-seq(min(plot1$x.lim),max(plot1$x.lim),abs(res))
 if(plot1$x.lim[1]>plot1$x.lim[2] & res>0){res<--res}
 plotx<-seq(plot1$x.lim[1],plot1$x.lim[2],res)}#set sequence of x values at which to plot, depending on plotting direction
 #plot spindles
-viol(plotx,pos=i, stat=stat, table=itable, smooth=smooth, dscale=dscale, col=col, fill=fill, lwd=lwd,lty=lty,cutoff=cutoff,w=w)
 
+if(exists("cutoff")){
+viol(plotx,pos=pos[i], stat=stat, table=itable, smooth=smooth, dscale=dscale, col=col, fill=fill, lwd=lwd,lty=lty,cutoff=cutoff,w=w)}else{
+viol(plotx,pos=pos[i], stat=stat, table=itable, smooth=smooth, dscale=dscale, col=col, fill=fill, lwd=lwd,lty=lty,w=w)} #plot with cutoff values if set, viol default if not
 
-}else if(is.matrix(occ) | is.data.frame(occ)){#if instead of a list object, a dataframe is given giving x and diversity values to plot (can also be co-opted to plot any other values, e.g. disparity
+}else if(is.matrix(occ) | is.data.frame(occ)){##if instead of a list object, a dataframe is given giving x and diversity values to plot (can also be co-opted to plot any other values, e.g. disparity
 if(!("x" %in% colnames(occ))){stop("if occ is a data.frame() or matrix(), it needs to have a column giving x values for plotting with column name == x")
 }
 if(taxsel[i] %in% colnames(occ)){
-
-viol(occ[,"x"], pos=i, stat=occ[,taxsel[i]], dscale=dscale, col=col, fill=fill, lwd=lwd,lty=lty, cutoff=cutoff, w=w)
+if(exists("cutoff")){
+viol(occ[,"x"], pos=pos[i], stat=occ[,taxsel[i]], dscale=dscale, col=col, fill=fill, lwd=lwd,lty=lty, cutoff=cutoff, w=w)}else{
+viol(occ[,"x"], pos=pos[i], stat=occ[,taxsel[i]], dscale=dscale, col=col, fill=fill, lwd=lwd,lty=lty, w=w)} #plot with cutoff values if set, viol default if not
 }
 
 }
@@ -1250,11 +1265,11 @@ if(length(txt.y)==1){txt.y<-rep(txt.y, length(taxsel))}
 if(length(txt.x)>1){#if a vector of x values for labels is provided
     if(length(which(names(txt.x)==taxsel[i]))==1){
     which(names(txt.x)==taxsel[i])->j
-    text(x=txt.x[j],y=i,adj=c(0,txt.y[i]), taxsel[i], cex=cex.txt,col=col.txt)
+    text(x=txt.x[j],y=pos[i],adj=c(adj.x[i],txt.y[i]), taxsel[i], cex=cex.txt,col=col.txt)
     }else{
-    text(x=txt.x[i],y=i,adj=c(0,txt.y[i]), taxsel[i], cex=cex.txt,col=col.txt)}
+    text(x=txt.x[i],y=pos[i],adj=c(adj.x[i],txt.y[i]), taxsel[i], cex=cex.txt,col=col.txt)}
 }else{
-text(x=txt.x,y=i,adj=c(0,txt.y[i]), taxsel[i], cex=cex.txt,col=col.txt)}
+text(x=txt.x,y=pos[i],adj=c(adj.x[i],txt.y[i]), taxsel[i], cex=cex.txt,col=col.txt)}
 }
 
 }#end loop